+++ block_conv.go
@@ -153,6 +153,8 @@ func (b *InputBlock) UnmarshalJSON(data []byte) error {
 		e = &FeedbackButtonsBlockElement{}
 	case "icon_button":
 		e = &IconButtonBlockElement{}
+	case "workflow_button":
+		e = &WorkflowButtonBlockElement{}
 	default:
 		return fmt.Errorf("unsupported block element type %v", s.TypeVal)
 	}
@@ -231,12 +233,18 @@ func (b *BlockElements) UnmarshalJSON(data []byte) error {
 			blockElement = &RadioButtonsBlockElement{}
 		case "static_select", "external_select", "users_select", "conversations_select", "channels_select":
 			blockElement = &SelectBlockElement{}
+		case "multi_static_select", "multi_external_select", "multi_users_select", "multi_conversations_select", "multi_channels_select":
+			blockElement = &MultiSelectBlockElement{}
 		case "number_input":
 			blockElement = &NumberInputBlockElement{}
+		case "file_input":
+			blockElement = &FileInputBlockElement{}
 		case "feedback_buttons":
 			blockElement = &FeedbackButtonsBlockElement{}
 		case "icon_button":
 			blockElement = &IconButtonBlockElement{}
+		case "workflow_button":
+			blockElement = &WorkflowButtonBlockElement{}
 		default:
 			return fmt.Errorf("unsupported block element type %v", blockElementType)
 		}
@@ -357,6 +365,12 @@ func (a *Accessory) UnmarshalJSON(data []byte) error {
 			return err
 		}
 		a.CheckboxGroupsBlockElement = element.(*CheckboxGroupsBlockElement)
+	case "workflow_button":
+		element, err := unmarshalBlockElement(r, &WorkflowButtonBlockElement{})
+		if err != nil {
+			return err
+		}
+		a.WorkflowButtonElement = element.(*WorkflowButtonBlockElement)
 	default:
 		element, err := unmarshalBlockElement(r, &UnknownBlockElement{})
 		if err != nil {
@@ -406,6 +420,12 @@ func toBlockElement(element *Accessory) BlockElement {
 	}
 	if element.MultiSelectElement != nil {
 		return element.MultiSelectElement
+	}
+	if element.RichTextInputElement != nil {
+		return element.RichTextInputElement
+	}
+	if element.WorkflowButtonElement != nil {
+		return element.WorkflowButtonElement
 	}
 
 	return nil